requirements.txt
flask==2.3.2
selenium==4.10.0
urllib3==1.26.16
pyOpenSSL==23.2.0
chromedriver-autoinstaller
flask-cors>=3.0.10
pytest==8.1.1
pytest-mock==3.14.0